parse removed enums moved up apimethod removed
