+++ .travis.yml
@@ -3,8 +3,6 @@ language: python
 sudo: false
 
 python:
-  - "3.2"
-  - "3.3"
   - "3.4"
   - "3.5"
