+++ betfairlightweight/__init__.py
@@ -1,6 +1,6 @@
 from .apiclient import APIClient
 from .exceptions import BetfairError
-from .streaming.listener import StreamListener
+from .streaming import StreamListener
 
 
 __version__ = '0.3.0'
